+++ src/main/java/com/medicationManagement/MedicationManagement/dto/EnderecoRequest.java
@@ -1,7 +1,6 @@
 package com.medicationManagement.MedicationManagement.dto;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
+
 import lombok.Data;
