+++ src/main/java/com/medicationManagement/MedicationManagement/dto/EstoqueResponse.java
@@ -3,15 +3,18 @@ package com.medicationManagement.MedicationManagement.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
-@AllArgsConstructor
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
 public class EstoqueResponse {
     private Long cnpj;
     private Integer nroRegistro;
     private Integer quantidade;
     private LocalDateTime dataAtualizacao;
 }
-// (DTO para a resposta de adição/atualização de medicamentos no estoque)
